adguard-dns-mcp/requirements.txt
fastmcp
dnspython
uvicorn
